llama-index-integrations/observability/llama-index-observability-otel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython==8.10.0",
    "jupyter>=1.0.0,<2",
    "mypy==0.991",
    "pre-commit==3.2.0",
    "pylint==2.15.10",
    "pytest==7.2.1",
    "pytest-mock==3.11.1",
    "ruff==0.0.292",
    "types-Deprecated>=0.1.0",
    "types-PyYAML>=6.0.12.12,<7",
    "types-protobuf>=4.24.0.4,<5",
    "types-redis==4.5.5.0",
    "types-requests==2.28.11.8",
    "types-setuptools==67.1.0.0",
    "black[jupyter]<=23.9.1,>=23.7.0",
    "codespell[toml]>=v2.2.6",
    "diff-cover>=9.2.0",
    "pytest-cov>=6.1.1",
]

[project]
name = "llama-index-observability-otel"
version = "0.1.0"
description = "llama-index observability integration with OpenTelemetry"
authors = [{name = "Clelia Astra Bertelli", email = "clelia@runllama.ai"}]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = "MIT"
dependencies = [
    "llama-index-core>=0.12.36,<0.13",
    "opentelemetry-sdk>=1.33.0,<2",
    "opentelemetry-api>=1.33.0,<2",
    "opentelemetry-semantic-conventions>=0.54b0,<1",
]

[tool.codespell]
check-filenames = true
check-hidden = true
ignore-words-list = "Gere"
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.hatch.build.targets.sdist]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.hatch.build.targets.wheel]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.llamahub]
contains_example = false
import_path = "llama_index.openinference.otel"

[tool.llamahub.class_authors]
OpenInferenceCallbackHandler = "llama-index"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.8"
